feat: Initial commit for the snekulator
The snekulator calculates the average number of moves
needed to complete a game of Snakes and Ladders,
specifically the Milton Bradley version of the game.

It sets up the transition matrix for the 100x100 and
incorporates the original 19 snakes and ladders.
It then proceeds to calculate the fundamental matrix,
which is used as the source of the solution.

For all matrix manipulation, the program heavily relies
on the GNU Scientific Library (GSL).